platform/xilinx: replace PLLs with DCMs on Spartan-6
Every clock management block in Spartan-6 has 2xDCM and 1xPLL. PLLs are therefore a scarcer resource
and should be preserved for user applications whenever possible. In this case, it is very much possible
to generate the necessary clocks via DCMs. Furthermore, we expose the unused CLKFX output of one of the
DCMs (along with a generic to configure the multiplication/division factors), in case it can be used
by the user application.