[wr_streamers_demo] Updated the streamers demo to work Boards Support Package
The streamers demo was using very old top. With the updated of resets,
etc, it stopped working (at least the testbench did stop). So, I finally
updated this top to work (in the testbench at least) with the new BSP.
This required a major re-do of the top. I left from the old as much
as I could. The new top is based on the spec_ref_design. This was
tested only for simulation
(testbech/wr_streamers/streamers-on-spec_trigger-distribution).
A commit with updates to simulation follows.