wr_streamers: included the generic to choose clk_data into streamers generic rx/tx records
Instead of adding another generic to streamers that later needs to
be propagated outside of Board Support Package and potentially for
each board, the choice of clock (clk_sys_i or clk_ref_i) is included
in the record of generics for tx/rx path of streamers. Indeed,
theoretically, such choice can be done per-path (rx and/or tx).