[CUTE] move oserdes_4_to_1 to board/cute (temporary solution)
oserdes_4_to_1.vhd is Xilinx-specific and cannot be in modules/
that include generic modules. This prevented the Altera-based
design from building. Ultimately, the module needs to be moved
to some more appropriate place, possibly to
general-cores/platform/xilinx. To be decided. A temporary solution
is to have it in board/cute, since it is only used by in the
xwrc_board_cute.