[streamers] re-arrange output diag/snmp array so it is backward-compatible
New statistics data was added to diag/snmp array. To ensure
backward-comapatibity between MIB versions, this new data needs
to be added at the end. Thus, now we need to inject dbg_word and
magic number in the middle.

The diag_ver is not bumped because the change is backward-compatible.