feat(debug) Added ILAs and VIOs for PLL/Clock dbg with dbg_component This commit is intended to add debug features to the kasli_soc design. It is not primarily intended to be pushed to main since some ports need to be added to the entities. At multicore-qc, the branch debug/kasli_wrpc depends on these files here to build the project with ILAs and VIOs and observe all clocks in the system.
At `xwrc_board_kasli.vhd`, the `clk_dbd_component` is instantiated as a hub of debug
associated signals. At some point, they can be connected to the dbg_bus_o, but not yet.
For each clock being sampled, there is a flop inverting the output `q_clk_name` such
that we are able to see the output fliping at the same frequency as the clock. The same
is done with counters, although the counters are there because one assumption was that
the gtp and pll_ref clocks could have not being sampled properly since all of them are
125MHz. It has proven to not be the case but counters can still be useful.

XDC constraints have been added at `clock.xdc` to `set_false_path` through the debug
cells, hence no CDC is being tackled here.