fix(i2c_slave+BRAM, i2c_eeprom, dbg) depends commit at multicore-qc
Depends on commit from multicore-qc (check notes since they are mutually dependent of
each other and this one was pushed first). A couple of core files and prim must be in
place before building project with fusesoc.

Again, it is not compatible with main branch.
There is a bunch of debug components that do not exist there.

In all pkg files, changed the base address for the storage memory.
It was 16#600000# (default for SVEC (M25P128)), and now is changed to 16#0# considering
there is no offset in EEPROM.

Debug registers: they are no longer replacing one wire.
The 1W peripheral is back and instead the debug registers are now connected to the space
that belonged to the wrpc usr diag reg (not the cpu).

At xwrc_board_kasli, a new i2c_wrapper_component is instantiated with a gc_i2c_slave,
BRAM, and ILA to replace the Kasli SoC EEPROM.